Move ViewSpecification to entities package
