+++ ldes-server-domain/src/main/java/be/vlaanderen/informatievlaanderen/ldes/server/domain/view/valueobject/ViewAddedEvent.java
@@ -1,7 +1,7 @@
 package be.vlaanderen.informatievlaanderen.ldes.server.domain.view.valueobject;
 
+import be.vlaanderen.informatievlaanderen.ldes.server.domain.viewcreation.entities.ViewSpecification;
 import be.vlaanderen.informatievlaanderen.ldes.server.domain.viewcreation.valueobjects.ViewName;
-import be.vlaanderen.informatievlaanderen.ldes.server.domain.viewcreation.valueobjects.ViewSpecification;
 
 public class ViewAddedEvent {
 	private final ViewSpecification viewSpecification;
